Completed implementation of Asif's Parameter classes. He allows the optimiser to work in a different space to the parameter values, so that the constraints required are maintained (i.e. kappa >=0, pi sum to 1)
This requires a mapping between these spaces. I've realised that the mapping
requires the implementation of another class, which houses the parameter instances.

In swmutsel, this is SwMut (and SwMutSel). My version is MutationModel,
which like SwMut descendes from an abstract base class called Substitution model.
This is effectively an HKY.
I'll later write a similar class which includes the selection parameters (the CAN)

These changes required some alterations to the TestModel class, in testFunction().
Passes all tests.

I also added default values for kappa and pi in Constants.java
and have other classes (CommandAegs and TsTvRatioAdvanced), which assign kappa,
refer to these defaults