I had been experiencing a very pernicious bug. CAN would mysteriously fail to compute same lnL as baseml for a given datasets, with the same paramter values. Consequently the optimisation was hopelessly unable to get correct param values.
Richard eventually suggested this was a scaling problem - and behold,
I had failed to scale the rate matrix Q such that the average sub rate = 1.

This scaling implemented, the correct lnL is now computed.

As this follows a lot of desperate debugging, there is a lot of stuff (e.g. print statements) to tidy up in this commit