Made -w flag arguments more flexible. Previously it was necessary to specify the initital w values for each gene, even though in an analysis you probably want these to be 1.0 every time.
Now the -w is optional, since the length of the relevent starting value array is worked out from the max of the gene numbers supplied with -a, -b and -c arguments. The default value of 1.0 is specified in Constants.java.

NB this does mean -w is also optional for Simulate program, but in worst case the user will accidentally use 1.0 for all genes (this might be good actually - allows you to simulate with HKY if you want to)

Also added new method to MatrixPrinter so it can print int[][] as well as double[][]