Following talk with Richard, we now realise we need to scale the sub rates by some factor nu such that the number of branch lengths match subs per site per bl
R's idea is to work out nu by simulation. So I am reimplementing evolveBranch
so that the code which performs the substitution is its own method,
so it can be called separately when evolving along a branch and when
computing nu

I've left the original body of code commented out in evolveBranch. will remove in next commit