had a long break while doing tests of the model with simulated data.
Found that the optimisation is performing poorly, in some cases with alt models fitting more poorly than null, and finding superior lnL values when calculating with the true parameter values used in the simulation

Richard recommends the BOBYQA opt algorithm implemented in Apache Commons Math, so I'm using that now.

In this commit I've written a new method in Optimise class which allows an optimisation to be performed with this new algorith.
Because the Apache docs are so bad we don't know what areguments you need to give it, so may need tweeking.
This code is untested